refactor: move mobile_restoreFromMnemonicWords
